Use different package name between restart/rolling bwc tests
