+++ qa/restart-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/restart/TextChunkingProcessorIT.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch.bwc;
+package org.opensearch.neuralsearch.bwc.restart;
 
 import java.net.URL;
 import java.nio.file.Files;
