+++ qa/rolling-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/rolling/TextChunkingProcessorIT.java
@@ -2,7 +2,7 @@
  * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.opensearch.neuralsearch.bwc;
+package org.opensearch.neuralsearch.bwc.rolling;
 
 import java.net.URL;
 import java.nio.file.Files;
